Remove spree_auth_devise as a dependency
This was in the original gem for Amazon but that had no tests. When
running the imported tests we got issues since devise isn't stubbed out.

Rather than stub out devise the easier solution is to do what the
solidus version of the gem did which is just remove the devise
dependency.

This gem mostly bypasses the Spree authentication anyway so no point
in testing with it enabled and then having to stub it out.