Stub actual amazon authorization
This isn't triggered on the Solidus one because the payment does not
actually have a payment type causing an early return. This new setup
is more representative of real data but it means it tries to actually
transact (as it should). To prevent it from hitting a real service it
is stubbed.