Remove version restrictions
The specs were written for Rails 5. For example the params was specified
as:

    post :payment, params: { some: 'param' }

instead of the old style:

    post :payment, some: 'param'

But the gemfile was bound to an older version of Spree/Rails/etc.
Removed those restrictions and did a `bundle update`.

Once we were running on the correct version of Rails the migrations
were not longer compatible. I adjusted them to be so. I'm unsure if we
will get issues on Magpie since it is on an older Rails but if so I will
just modify the imported migrations.