Don't validate the authenticity token
Previously this worked because Rails used to be bundled with jquery_ujs
to implement the unintrusive javascript. A side effect of this library
is that it automatically includes the authenticity token with every
jquery-based XHR request[1].

This means the XHR request included the token. But with newer rails the
dependency on jQuery has been removed and UJS is now implemented via
vanilla JS. This now means that jquery-based XHR requests are no longer
automatically include the token causing these requests to fail.

We could potentially include the token with this specific XHR request
but simpler is just to not require the token. The security risk is
almost non-existant.

If this commit were applied to the solidus branch this would close
boomerdigital/solidus_amazon_payments#54.

1. https://github.com/rails/jquery-ujs/blob/master/src/rails.js#L69-L73