Remove bad error detection
There error detection doesn't really do the right thing here. These
pages are called via XHR requests so a traditional HTTP redirect won't
work. The actual end result is that the page it redirects to (whole page
including layout) gets placed inside a container (causing a full page to
be inside a page).

These errors are triggered often due to race coditions. The order is
ONLY assigned when submitting to create the payment object but the
other pages are often called earlier.

To avoid the issues I am just changing this to silently ignore errors.
This is probably not the correct solution. In general this controller
isn't what I would call best practices. But sort of refactoring this
into being something decent this seems like the best solution.