Add payment differentiator
The amazon order number is assigned as the payment number. The problem
is that this number needs to be unique and it is possible we can get
more than one payment in the same order. For example:

1. User starts amazon payment
2. User then switches to regular payment
3. User finally switches back to amazon payment

The regular payment invalidated the first amazon payment. The system
correctly then creates a new amazon payment but the uniqueness
constraint causes an error.

This actually happens more often than you would think as payments are
often validated as carts are merged/etc.

Adding a number to the end of the payment corrects this problem as it
ensure each payment number is unique.

I did attempt to use the payment id. But this is fed into some sort of
authorization request ID that cannot be more than 32 characters. The
payment id makes it exceed this easily.

Also the use of `_` was purposeful. The only characters allowed in
the authorization request ID is a-zA-Z0-9_-. The `-` is used for some
sort of denormalized value so to avoid problems I used `_`.