Make login partial more embeddable
If calling just:

    <%= render 'spree/amazon/login' %>

this behaves exactly as before. but we may want to use this same blurb
of code to handle other variations of Amazon interaction (such as just
login). To do this we allow the `button_type` to be set and the
`return_path` to be set.

Finally rather than assume @order in the current context we use the
`current_order` helper (creating an order if neccessary).