Skip default credit card assignment on amazon payment
Solidus populates a dummy payment record with the default credit card
if there is one. But this invalidates the amazon payment causing an
error when confirming the order.

To avoid this this assignment is skipped IF there is a valid amazon
payment.

This will work out OK if the user changes their mind and does a normal
checkout as that will invalidate the amazon payment.